azurecloud/eval/requirements.txt
# RAG Evaluation Framework Dependencies

# Azure SDK
azure-identity>=1.15.0
azure-ai-evaluation>=1.0.0
azure-ai-projects>=1.0.0
azure-cosmos>=4.5.0
azure-search-documents>=11.4.0

# OpenAI
openai>=1.12.0

# HTTP
httpx>=0.26.0

# ML/Data
scikit-learn>=1.4.0
numpy>=1.26.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Utilities
python-dotenv>=1.0.0
